* Added framework so we can research captured/killed aliens soon. * This means i changed the boolean "statusColelcted to an int.

+++ src/client/cl_research.h
@@ -75,7 +75,7 @@ typedef struct technology_s
 	char	mdl_top[MAX_QPATH];
 	char	mdl_bottom[MAX_QPATH];
 
-	byte	statusCollected;				// Did we loot this item?
+	int	statusCollected;				// Did we loot this item (and how mach of it -> aliens/corpses) ?
 	byte	needsCollected;				// Is a collected item neccessary to research this item?
 	byte	statusResearchable;			// Is this item researchable?
